installing packages
We made a list of packages we wanted to install in the pkgs list. Then we ran a for loop to check that every packaged you needed was in the list. Within the for loop, we say "if p is not in pkgs, install the package"

This makes it easy for anyone to recieve your file and have it run, even if they do not already have the package.